+++ src/test/kotlin/snc/openchargingnetwork/node/integration/IntegrationTest.kt
@@ -233,4 +233,5 @@ class IntegrationTest {
             sleep(1000)
         }
     }
+
 }