+++ src/test/kotlin/snc/openchargingnetwork/node/integration/parties/PartyServer.kt
@@ -12,6 +12,7 @@ import snc.openchargingnetwork.node.models.OcnServicePermission
 import snc.openchargingnetwork.node.models.OcnRulesListType
 import snc.openchargingnetwork.node.models.ocpi.*
 import snc.openchargingnetwork.node.tools.generateUUIDv4Token
+import snc.openchargingnetwork.node.tools.toBs64String
 
 open class PartyServer(val config: PartyDefinition, deployedContracts: OcnContracts) {
 
@@ -36,7 +37,7 @@ open class PartyServer(val config: PartyDefinition, deployedContracts: OcnContra
         }
 
         app.before {
-            if (it.header("Authorization") != "Token $tokenB") {
+            if (it.header("Authorization") != "Token ${tokenB.toBs64String()}") {
                 throw JavalinException(message = "Unauthorized")
             }
         }
@@ -81,7 +82,8 @@ open class PartyServer(val config: PartyDefinition, deployedContracts: OcnContra
 
     fun registerCredentials() {
         // TODO: could also request versions and store endpoints in memory
-        val tokenA = getTokenA(node, listOf(config.party))
+        val tokenA = getTokenA(node, listOf(config.party)).toBs64String()
+
         val response = khttp.post("$node/ocpi/2.2/credentials",
                 headers = mapOf("Authorization" to "Token $tokenA"),
                 json = coerceToJson(Credentials(
@@ -92,7 +94,12 @@ open class PartyServer(val config: PartyDefinition, deployedContracts: OcnContra
                                 businessDetails = BusinessDetails(name = "Some CPO"),
                                 countryCode = config.party.country,
                                 partyID = config.party.id)))))
-        tokenC = response.jsonObject.getJSONObject("data").getString("token")
+
+        tokenC = response
+            .jsonObject
+            .getJSONObject("data")
+            .getString("token")
+            .let { it.toBs64String() }
     }
 
     fun deleteCredentials() {